fix: harden companion reliability (crash recovery, dead-input)
Reliability hardening so a transient fault no longer leaves controller
input permanently dead until manual restart/reboot:

- install.rs: configure SCM failure actions (sc failure) so the service
  auto-restarts 60s after a crash.
- xinput_backend.rs secure poll thread:
  - catch_unwind inside the extern "system" wndproc (a panic unwinding
    across the FFI boundary aborts the whole process on Rust >=1.81).
  - supervisor detects rx disconnect and respawns the dead thread.
  - WM_POWERBROADCAST resume handler flushes stale XUSB/HID handles.
  - replace blocking GetMessageW with MsgWaitForMultipleObjects(250ms)
    + PeekMessage drain; >750ms silence exits for respawn.
  - throttle XInputGetState error logs (was ~125/s -> disk fill).
  - HID reconnect re-enum window 1s -> 150ms when no readers.
- pipe_server.rs: wrap LED engine loop body in catch_unwind.

Companion stays app-agnostic (no client coupling added).
cargo check clean; 92 tests pass.

Note: this branch's working tree also carried unrelated in-progress
changes across ~20 other files predating this work; committed together
per request. .lavish review artifact gitignored.

+++ src/win/desktop.rs
@@ -119,7 +119,7 @@ unsafe fn desktop_name(h: HDESK) -> Option<String> {
     let mut needed = 0u32;
     GetUserObjectInformationW(
         HANDLE(h.0),
-        USER_OBJECT_INFORMATION_INDEX(UOI_NAME.0 as i32),
+        USER_OBJECT_INFORMATION_INDEX(UOI_NAME.0),
         Some(buf.as_mut_ptr().cast()),
         (buf.len() * 2) as u32,
         Some(&mut needed),
